+++ packages/file-storage/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This is the changelog for [`file-storage`](https://github.com/remix-run/remix/tree/main/packages/file-storage). It follows [semantic versioning](https://semver.org/).
 
+## Unreleased
+
+- BREAKING CHANGE: `MemoryFileStorage` class has been replaced with a `createMemoryFileStorage()` factory function. Update your code from `new MemoryFileStorage()` to `createMemoryFileStorage()`.
+- BREAKING CHANGE: `LocalFileStorage` class has been replaced with a `createLocalFileStorage(directory)` factory function. Update your code from `new LocalFileStorage(directory)` to `createLocalFileStorage(directory)`.
+
 ## v0.12.0 (2025-11-20)
 
 - Add `@remix-run/fs` as a peer dependency. This package now imports from `@remix-run/fs` instead of `@remix-run/lazy-file/fs`.